agent_engine/blog_generator/requirements.txt
pydantic==2.12.3
pydantic-settings==2.11.0
python-dotenv==1.2.1
openai==2.6.1
openai-agents==0.4.2
bs4
aiohttp
python-frontmatter